+++ finetrainers/patches/__init__.py
@@ -11,6 +11,7 @@ if TYPE_CHECKING:
 def perform_patches_for_training(args: "BaseArgs", parallel_backend: "ParallelBackendType") -> None:
     # To avoid circular imports
     from finetrainers.config import ModelType, TrainingType
+
     from .dependencies.diffusers import patch
 
     patch.patch_diffusers_rms_norm_forward()
