+++ finetrainers/typing.py
@@ -1,4 +1,3 @@
-from enum import Enum
 from typing import Union
 
 from diffusers import CogVideoXDDIMScheduler, FlowMatchEulerDiscreteScheduler
@@ -10,23 +9,3 @@ from .data import ImageArtifact, VideoArtifact
 ArtifactType = Union[ImageArtifact, VideoArtifact]
 SchedulerType = Union[CogVideoXDDIMScheduler, FlowMatchEulerDiscreteScheduler]
 TokenizerType = Union[CLIPTokenizer, T5Tokenizer, T5TokenizerFast, LlamaTokenizer, LlamaTokenizerFast]
-
-
-class ControlType(str, Enum):
-    r"""
-    Enum class for the control types.
-    """
-
-    CANNY = "canny"
-    CUSTOM = "custom"
-
-
-class FrameConditioningType(str, Enum):
-    r"""
-    Enum class for the frame conditioning types.
-    """
-
-    INDEX = "index"
-    PREFIX = "prefix"
-    RANDOM = "random"
-    SUFFIX = "suffix"
